flask_app/requirement.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
attrs==23.2.0
Automat==22.10.0
babel==2.17.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
comm==0.2.2
constantly==23.10.4
construct==2.10.70
contourpy==1.2.0
cssselect==1.2.0
cv==1.0.0
cycler==0.12.1
cytoolz==0.12.3
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnslib==0.9.24
dnspython==2.7.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
executing==2.0.1
fake-headers==1.0.2
fake-useragent==1.4.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.0
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
graphviz==0.20.3
groq==0.22.0
grpcio==1.65.2
grpcio-status==1.65.2
h11==0.14.0
hedera-sdk-py==2.29.0
hexbytes==0.3.1
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.21.4
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.3.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-groq==0.3.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.32
loguru==0.7.2
lru-dict==1.2.0
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
MechanicalSoup==1.3.0
mistune==3.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.35.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opacus==1.4.1
openai==1.75.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parse==1.20.1
parsimonious==0.9.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.23.0
ply==3.11
prometheus_client==0.21.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==8.2.2
Pygments==2.17.2
pyjnius==1.6.1
pyparsing==3.1.2
pyppeteer==1.0.2
pyquery==2.0.0
pysmi==0.3.4
pysnmp==4.4.12
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.2.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scapy==2.5.0
scipy==1.13.0
Scweet==1.8
selenium==4.17.2
Send2Trash==1.8.3
setuptools==69.1.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.44.1
sympy==1.12
tenacity==9.0.0
tensor==0.3.6
terminado==0.18.1
tesseract==0.1.3
textblob==0.18.0.post0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
trio==0.24.0
trio-websocket==0.11.1
tweepy==4.14.0
Twisted==24.3.0
twitter-scraper==0.4.4
twscrape==0.11.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
w3lib==2.1.2
wcwidth==0.2.13
web3==6.15.1
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
zstandard==0.23.0
